refactor home fragment
