+++ DDUBUCK/app/src/main/AndroidManifest.xml
@@ -42,10 +42,6 @@
             </intent-filter>
         </activity>
 
-
-        <activity android:name=".home.HomeActivity">
-        </activity>
-
         <meta-data
             android:name="com.naver.maps.map.CLIENT_ID"
             android:value="lz0an3yhfn" />
