+++ DDUBUCK/app/src/main/java/com/example/ddubuck/data/data.kt
@@ -1,0 +1,3 @@
+package com.example.ddubuck.data
+
+//data management file
