+++ DDUBUCK/app/src/main/java/com/example/ddubuck/ui/home/bottomSheet/BottomSheetFreeProgressFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ddubuck.home.bottomSheet
+package com.example.ddubuck.ui.home.bottomSheet
 
 import android.os.Bundle
 import android.util.Log
@@ -9,10 +9,10 @@ import android.widget.Button
 import androidx.fragment.app.Fragment
 import com.example.ddubuck.R
 
-class BottomSheetCourseProgressFragment : Fragment() {
+class BottomSheetFreeProgressFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val rootView = inflater.inflate(R.layout.bottom_sheet_course_progress, container, false)
+        val rootView = inflater.inflate(R.layout.bottom_sheet_free_progress, container, false)
         val pauseButton : Button = rootView.findViewById(R.id.sheet_free_progress_pause)
         val endButton : Button = rootView.findViewById(R.id.sheet_free_progress_end)
         return rootView
