+++ DDUBUCK/app/src/main/java/com/example/ddubuck/ui/home/bottomSheet/BottomSheetSelectRvAdapter.kt
@@ -1,19 +1,20 @@
-package com.example.ddubuck.home
+package com.example.ddubuck.ui.home.bottomSheet
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.ddubuck.R
-import com.example.ddubuck.home.bottomSheet.BottomSheetSelectFragment
-import com.naver.maps.map.e
-import kotlin.math.log
+import com.example.ddubuck.data.home.CourseItem
+import com.example.ddubuck.data.home.WalkRecord
+import com.example.ddubuck.ui.home.bottomSheet.BottomSheetCourseDetailFragment
+import com.example.ddubuck.ui.home.bottomSheet.BottomSheetFreeDetailFragment
 
 class HomeRvAdapter(private val itemList: ArrayList<CourseItem>,
-                    private val callback: BottomSheetSelectFragment.OnCourseSelectedListener,):
+                    private val fm: FragmentManager,):
     RecyclerView.Adapter<Holder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
@@ -22,7 +23,7 @@ class HomeRvAdapter(private val itemList: ArrayList<CourseItem>,
     }
 
     override fun onBindViewHolder(holder: Holder, position: Int) {
-        holder.bind(itemList[position], callback)
+        holder.bind(itemList[position], fm)
     }
 
     override fun getItemCount(): Int {
@@ -35,9 +36,10 @@ class Holder(itemView: View?) : RecyclerView.ViewHolder(itemView!!)  {
     private val body = itemView?.findViewById<TextView>(R.id.sheet_item_body)
     private val picture = itemView?.findViewById<ImageView>(R.id.sheet_item_picture)
 
-    fun bind(i: CourseItem,callback : BottomSheetSelectFragment.OnCourseSelectedListener) {
+    fun bind(i: CourseItem, fm: FragmentManager) {
         if(i.isFreeWalk) {
-            itemView.setOnClickListener{callback.onCourseSelected(CourseItem(
+            itemView.setOnClickListener{selectItem(fm,
+                    CourseItem(
                 true,
                 "자유산책",
                 "자유산책입니다",
@@ -46,10 +48,29 @@ class Holder(itemView: View?) : RecyclerView.ViewHolder(itemView!!)  {
             body?.text = "나만의 자유로운 산책,\n즐길 준비 되었나요?"
             picture?.setImageResource(R.mipmap.ic_launcher)
         } else {
-            itemView.setOnClickListener{callback.onCourseSelected(i)}
+            itemView.setOnClickListener{selectItem(fm,i)}
             title?.text = i.title
             body?.text = "걸음수 : ${i.walkRecord.stepCount}\n거리 : ${i.walkRecord.distance}\n시간 : ${i.walkRecord.walkTime}"
             picture?.setImageResource(R.mipmap.ic_launcher)
         }
+    }
+
+    fun selectItem(fm:FragmentManager, courseItem : CourseItem) {
+        if(courseItem.isFreeWalk) {
+            val frag = BottomSheetFreeDetailFragment()
+            val fmTransaction = fm.beginTransaction()
+            fmTransaction.setCustomAnimations(R.anim.fragment_fade_enter, R.anim.fragment_fade_exit)
+            fmTransaction.replace(R.id.bottom_sheet_container,frag, BOTTOM_SHEET_CONTAINER_TAG).addToBackStack(DETAIL_PAGE_FRAG).commit()
+        } else {
+            val frag = BottomSheetCourseDetailFragment(courseItem)
+            val fmTransaction = fm.beginTransaction()
+            fmTransaction.setCustomAnimations(R.anim.fragment_fade_enter, R.anim.fragment_fade_exit)
+            fmTransaction.replace(R.id.bottom_sheet_container,frag, BOTTOM_SHEET_CONTAINER_TAG).addToBackStack(DETAIL_PAGE_FRAG).commit()
+        }
+    }
+
+    companion object {
+        private const val BOTTOM_SHEET_CONTAINER_TAG = "BOTTOM_SHEET_CONTAINER"
+        private const val DETAIL_PAGE_FRAG = "DETAIL_PAGE_FRAG"
     }
 }
