--- DDUBUCK/app/src/main/res/layout/activity_home.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".home.HomeActivity">
-    <FrameLayout
-        android:id="@+id/home_map_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-    <FrameLayout
-        android:id="@+id/home_weather_score"
-        android:layout_marginTop="17dp"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginBottom="34dp"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"/>
-    <include layout="@layout/bottom_sheet"
-        android:id="@+id/home_bottom_sheet"
-        android:background="@android:color/white" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
