+++ DDUBUCK/app/src/main/res/layout/fragment_home.xml
@@ -1,24 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.coordinatorlayout.widget.CoordinatorLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment">
-
-
-
-    <TextView
-        android:id="@+id/text_home"
+    android:layout_height="match_parent">
+    <FrameLayout
+        android:id="@+id/home_map_container"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+        android:layout_height="match_parent"/>
+    <FrameLayout
+        android:id="@+id/home_weather_score"
+        android:layout_marginTop="17dp"
+        android:layout_marginHorizontal="16dp"
+        android:layout_marginBottom="34dp"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"/>
+    <include layout="@layout/bottom_sheet"
+        android:id="@+id/home_bottom_sheet"
+        android:background="@android:color/white" />
+
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
