--- DDUBUCK/app/src/main/res/navigation/mobile_navigation.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_home">
-
-    <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.example.ddubuck.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/navigation_challenge"
-        android:name="com.example.ddubuck.ui.challenge.ChallengeFragment"
-        android:label="@string/title_challenge"
-        tools:layout="@layout/fragment_challenge" />
-
-    <fragment
-        android:id="@+id/navigation_badge"
-        android:name="com.example.ddubuck.ui.badge.BadgeFragment"
-        android:label="@string/title_badge"
-        tools:layout="@layout/fragment_badge" />
-    <fragment
-        android:id="@+id/navigation_mypage"
-        android:name="com.example.ddubuck.ui.mypage.MyPageFragment"
-        android:label="@string/title_mypage"
-        tools:layout="@layout/fragment_mypage" />
-</navigation>
